pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
phytclust = ["assets/*"]

[project]
name = "phytclust"
version = "0.1.1"
description = "PhytClust takes a phylogenetic tree as input and outputs monophyletic leaf clusters that best fit the tree structure."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Katyayni Ganesan", email = "katyayni.ganesan@iccb-cologne.org" },
    { name = "Roland F Schwarz", email = "roland.schwarz@iccb-cologne.org" },
]
maintainers = [
    { name = "Katyayni Ganesan", email = "katyayni.ganesan@iccb-cologne.org" },
]
keywords = [
    "phylogenetics",
    "clustering",
    "taxonomy",
    "classification",
    "dynamic programming",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
]

requires-python = ">=3.10"

dependencies = [
    "numpy>=1.24",
    "biopython>=1.85",
    "matplotlib>=3.7",
    "pandas>=2.0",
    "scipy>=1.10",
    "PyYAML>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=24.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.0",
]
gui = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
]
docs = ["mkdocs>=1.6,<2", "mkdocs-material>=9.5,<10"]

[project.urls]
Documentation = "https://iccb-cologne.github.io/PhytClust/"
Source = "https://github.com/ICCB-Cologne/PhytClust"
Tracker = "https://github.com/ICCB-Cologne/PhytClust/issues"

[project.scripts]
phytclust = "phytclust.cli:main"
